requirements.txt
asgiref==3.8.1
Django==5.2
django-restframework==0.0.1
djangorestframework==3.16.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
